refactor: move out key / name resolving into data utils. cleans up smolvla
